refactor(reserve): modularize views and update barrel
