+++ mobile/rupu/lib/presentation/views/reserve/widgets/contact_row.dart
@@ -1,0 +1,41 @@
+import 'package:flutter/material.dart';
+
+class ContactRow extends StatelessWidget {
+  const ContactRow({
+    super.key,
+    required this.icon,
+    required this.label,
+    required this.value,
+  });
+
+  final IconData icon;
+  final String label;
+  final String value;
+
+  @override
+  Widget build(BuildContext context) {
+    final cs = Theme.of(context).colorScheme;
+    final tt = Theme.of(context).textTheme;
+    return Padding(
+      padding: const EdgeInsets.only(bottom: 10),
+      child: Row(
+        children: [
+          Icon(icon, color: cs.onSurfaceVariant),
+          const SizedBox(width: 10),
+          Text(
+            label,
+            style: tt.labelMedium!.copyWith(color: cs.onSurfaceVariant),
+          ),
+          const SizedBox(width: 6),
+          Expanded(
+            child: Text(
+              value,
+              textAlign: TextAlign.end,
+              style: tt.bodyMedium!.copyWith(fontWeight: FontWeight.w700),
+            ),
+          ),
+        ],
+      ),
+    );
+  }
+}
